+++ Source/DotSpatial.Plugins.WebMap/WebMapPlugin.cs
@@ -301,6 +301,9 @@ namespace DotSpatial.Plugins.WebMap
                 // reproject view extents
                 Reproject.ReprojectPoints(viewExtentXy, viewExtentZ, App.Map.Projection, _webMercProj, 0, 2);
 
+                // fix the extents if they are invalid
+                viewExtentXy = GetValidExtent(viewExtentXy);
+
                 // set the new map extents
                 App.Map.ViewExtents = new Extent(viewExtentXy);
 
@@ -355,6 +358,37 @@ namespace DotSpatial.Plugins.WebMap
             _featureSetLayer.LegendItemVisible = false;
 
             App.Map.ZoomToMaxExtent();
+        }
+
+        private double[] GetValidExtent(double[] viewExtentXy)
+        {
+            // determine if any coordinate of the reprojected extents are valid
+            int idx = viewExtentXy.ToList().FindIndex(d => double.IsNaN(d) || double.IsInfinity(d));
+            if (idx == -1)
+            {
+                // all coordinates in the extent are valid so return the extent as-is
+                return viewExtentXy;
+            }
+            else
+            {
+                // one or more coordinates are invalid so try to guess a valid value
+                // assume the coordinates are in this order minX, minY, maxX, maxY
+                double minX = viewExtentXy[0];
+                if (double.IsNaN(minX) || double.IsInfinity(minX)) minX = TileCalculator.MinWebMercX;
+
+                double minY = viewExtentXy[1];
+                if (double.IsNaN(minY) || double.IsInfinity(minY)) minY = TileCalculator.MinWebMercY;
+
+                double maxX = viewExtentXy[2];
+                if (double.IsNaN(maxX) || double.IsInfinity(maxX)) maxX = TileCalculator.MaxWebMercX;
+
+                double maxY = viewExtentXy[3];
+                if (double.IsNaN(maxY) || double.IsInfinity(maxY)) maxY = TileCalculator.MaxWebMercY;
+
+                double[] viewExtNew = { minX, minY, maxX, maxY };
+
+                return viewExtNew;
+            }
         }
 
         private bool InsertBaseMapLayer(IMapGroup group)