+++ Changelog.md
@@ -159,3 +159,4 @@ Be aware that code written for 1.9 will not work out of the box because DotSpati
 - Potential Bug in RasterBoundsExt class CellsContainingExtent(...) method (#1332)
 - Potential bug in EnvelopeExt (and ExtentExt) class Reproportion(...) method (#1326)
 - Bug in AzimuthalEquidistant class (#1342)
+- Bug in extent calculation in WebMap plugin (#)